CLI-60: Refactor env config file processing
