+++ lib/CLIManager.js
@@ -56,6 +56,7 @@ const EnvironmentsService = require('./environment/EnvironmentsService');
 const CollectionsService = require('./collection/CollectionsService');
 const BusinessLogicService = require('./BusinessLogicService');
 
+const EnvFileProcessor = require('./EnvFileProcessor');
 const ConfigFileProcessor = require('./ConfigFileProcessor');
 
 /**
@@ -95,14 +96,14 @@ class CLIManager {
     this._controllers[Namespace.COLL] = new CollectionsController({ cliManager: this, collectionsService, environmentsService });
 
     const businessLogicService = new BusinessLogicService(this);
-    this._configFileProcessor = new ConfigFileProcessor({
+    const envFileProcessor = new EnvFileProcessor({
       cliManager: this,
-      organizationsService,
-      applicationsService,
       environmentsService,
       collectionsService,
       businessLogicService
     });
+
+    this._configFileProcessor = new ConfigFileProcessor({ envFileProcessor });
   }
 
   getController(name) {
